# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Course 后端 API 服务",
        "title": "AI Course API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/submission/assignment/{assignment_id}": {
            "get": {
                "description": "获取学生特定作业的详细信息和提交状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业提交"
                ],
                "summary": "获取学生特定作业详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "assignment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/submission/draft": {
            "post": {
                "description": "学生保存作业答案为草稿",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业提交"
                ],
                "summary": "保存作业草稿",
                "parameters": [
                    {
                        "description": "提交信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/submission/student/assignments": {
            "get": {
                "description": "获取当前学生的所有作业列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业提交"
                ],
                "summary": "获取学生作业列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/submission/submit": {
            "post": {
                "description": "学生提交作业",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业提交"
                ],
                "summary": "提交作业",
                "parameters": [
                    {
                        "description": "提交信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/submission/{id}": {
            "get": {
                "description": "获取提交的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业提交"
                ],
                "summary": "获取提交详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "提交ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "提交不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AnswerRequest": {
            "type": "object",
            "required": [
                "content",
                "question_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.SubmissionRequest": {
            "type": "object",
            "required": [
                "answers",
                "assignment_id"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AnswerRequest"
                    }
                },
                "assignment_id": {
                    "type": "integer"
                },
                "status": {
                    "enum": [
                        "draft",
                        "submitted"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SubmissionStatus"
                        }
                    ]
                }
            }
        },
        "model.SubmissionStatus": {
            "type": "string",
            "enum": [
                "draft",
                "submitted",
                "graded"
            ],
            "x-enum-comments": {
                "SubmissionStatusDraft": "草稿",
                "SubmissionStatusGraded": "已批改",
                "SubmissionStatusSubmitted": "已提交"
            },
            "x-enum-varnames": [
                "SubmissionStatusDraft",
                "SubmissionStatusSubmitted",
                "SubmissionStatusGraded"
            ]
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}